Refactor VLM inference microservice to VLM OV serving with updated structure
