--- microservices/vlm-inference-microservice/.gitignore
@@ -1,31 +1,0 @@
-# Ignore environment variables
-.env
-
-# Ignore Python cache files
-__pycache__/
-*.py[cod]
-
-# Ignore virtual environment directories
-venv/
-env/
-
-# Ignore log files
-*.log
-
-# Ignore IDE specific files
-.vscode/
-
-# Ignore model files
-ov-model/
-
-# Ignore pytest cache
-.pytest_cache/
-
-# Ignore pytest result files
-*.pytest_cache
-
-# pytest related dir
-empty_model_dir
-mock_model_dir
-.coverage
-htmlcov
