--- microservices/vlm-inference-microservice/compose.gpu.yaml
@@ -1,41 +1,0 @@
-services:
-  intel-egai-vlm-inference-microservice:
-    image: ${REGISTRY:-}intel-egai-vlm-inference-microservice-gpu:${TAG:-latest}
-    container_name: vlm-inference-microservice
-    build:
-      context: .
-      dockerfile: docker/Dockerfile.gpu
-    ports:
-      - "${VLM_SERVICE_PORT}:8000"
-    ipc: host
-    environment:
-      no_proxy_env: ${no_proxy_env},${HOST_IP}
-      no_proxy: ${no_proxy_env},${HOST_IP}
-      http_proxy: ${http_proxy}
-      https_proxy: ${https_proxy}
-      VLM_MODEL_NAME: ${VLM_MODEL_NAME}
-      VLM_COMPRESSION_WEIGHT_FORMAT: ${VLM_COMPRESSION_WEIGHT_FORMAT}
-      VLM_DEVICE: ${VLM_DEVICE}
-      VLM_SEED: ${VLM_SEED}
-      WORKERS: ${WORKERS:-1}
-    restart: unless-stopped
-    devices:
-      - /dev/dri:/dev/dri
-    healthcheck:
-      test: ["CMD-SHELL", "curl -f http://localhost:8000/health || exit 1"]
-      interval: 10s
-      timeout: 10s
-      retries: 40
-      start_period: 10s
-    networks:
-      - my_network
-    volumes:
-      - "/home/$USER/.cache/huggingface:/root/.cache/huggingface"
-      - ov-models:/app/ov-model
-volumes:
-  ov-models:
-    external: true
-
-networks:
-  my_network:
-    driver: bridge
