--- microservices/vlm-inference-microservice/logging.conf
@@ -1,34 +1,0 @@
-
-[loggers]
-keys=root,gunicorn,uvicorn
-
-[handlers]
-keys=console
-
-[formatters]
-keys=generic
-
-[logger_root]
-level=INFO
-handlers=console
-
-[logger_gunicorn]
-level=INFO
-handlers=console
-propagate=0
-qualname=gunicorn
-
-[logger_uvicorn]
-level=INFO
-handlers=console
-propagate=0
-qualname=uvicorn
-
-[handler_console]
-class=StreamHandler
-formatter=generic
-args=(sys.stdout,)
-
-[formatter_generic]
-format=%(asctime)s [%(process)d] [%(levelname)s] %(message)s
-datefmt=%Y-%m-%d %H:%M:%S
