--- microservices/vlm-inference-microservice/tests/__init__.py
@@ -1,2 +1,0 @@
-
-# This file is intentionally left blank.
