--- microservices/vlm-inference-microservice/tests/pytest.ini
@@ -1,6 +1,0 @@
-
-[pytest]
-testpaths = .
-python_files = test_*.py
-filterwarnings =
-    ignore:The `openvino.runtime` module is deprecated:DeprecationWarning
