+++ microservices/vlm-ov-serving/compose.yaml
@@ -1,0 +1,42 @@
+# Copyright (C) 2025 Intel Corporation
+# SPDX-License-Identifier: Apache-2.0
+
+services:
+  vlm-ov-serving:
+    image: ${REGISTRY:-}vlm-ov-serving:${TAG:-latest}
+    container_name: vlm-ov-serving
+    build:
+      context: .
+      dockerfile: docker/Dockerfile
+    ports:
+      - "${VLM_SERVICE_PORT}:8000"
+    ipc: host
+    environment:
+      no_proxy_env: ${no_proxy_env},${HOST_IP}
+      no_proxy: ${no_proxy_env},${HOST_IP}
+      http_proxy: ${http_proxy}
+      https_proxy: ${https_proxy}
+      VLM_MODEL_NAME: ${VLM_MODEL_NAME}
+      VLM_COMPRESSION_WEIGHT_FORMAT: ${VLM_COMPRESSION_WEIGHT_FORMAT}
+      VLM_DEVICE: ${VLM_DEVICE}
+      VLM_SEED: ${VLM_SEED}
+      WORKERS: ${WORKERS:-1}
+    restart: unless-stopped
+    healthcheck:
+      test: ["CMD-SHELL", "curl -f http://localhost:8000/health || exit 1"]
+      interval: 10s
+      timeout: 10s
+      retries: 40
+      start_period: 10s
+    networks:
+      - my_network
+    volumes:
+      - "/home/$USER/.cache/huggingface:/root/.cache/huggingface"
+      - ov-models:/app/ov-model
+volumes:
+  ov-models:
+    external: true
+
+networks:
+  my_network:
+    driver: bridge
