+++ microservices/vlm-ov-serving/src/config/model_config.yaml
@@ -1,0 +1,24 @@
+# Copyright (C) 2025 Intel Corporation
+# SPDX-License-Identifier: Apache-2.0
+
+qwen2.5-vl-7b-instruct:
+# The default range for the number of visual tokens per image in the model is 4-16384.
+# You can set min_pixels and max_pixels according to your needs, such as a token range of 256-1280, to balance performance and cost.
+  min_pixels: 128 * 28 * 28
+  max_pixels: 256 * 256 * 256
+
+qwen2.5-vl-3b-instruct:
+  min_pixels: 256 * 28 * 28
+  max_pixels: 256 * 28 * 28
+
+qwen2-vl-2b-instruct:
+  min_pixels: 256 * 28 * 28
+  max_pixels: 256 * 28 * 28
+
+phi-3.5-vision-instruct:
+  min_pixels: null
+  max_pixels: null
+
+phi-3-vision-128k-instruct:
+  min_pixels: null
+  max_pixels: null
