+++ microservices/vlm-ov-serving/tests/pytest.ini
@@ -1,0 +1,9 @@
+# Copyright (C) 2025 Intel Corporation
+# SPDX-License-Identifier: Apache-2.0
+
+
+[pytest]
+testpaths = .
+python_files = test_*.py
+filterwarnings =
+    ignore:The `openvino.runtime` module is deprecated:DeprecationWarning
